+++ R/map.R
@@ -217,7 +217,7 @@ walk3 <- function(.x, .y, .z, .f, ...) {
 #' @rdname map2
 walk_n <- function(.l, .f, ...) {
   .f <- as_function(.f)
-  args_list <- recycle_args(.l) %>% zip_n()
+  args_list <- recycle_args(.l) %>% transpose()
   for (args in args_list) {
     do.call(".f", c(args, list(...)))
   }
@@ -244,9 +244,9 @@ walk_n <- function(.l, .f, ...) {
 #' @name conditional-map
 #' @examples
 #' list(x = rbernoulli(100), y = 1:100) %>%
-#'   zip_n() %>%
+#'   transpose() %>%
 #'   map_if("x", ~ update_list(., y = ~ y * 100)) %>%
-#'   zip_n(.simplify = TRUE)
+#'   transpose(.simplify = TRUE)
 #'
 #' # Convert factors to characters
 #' iris %>%